examples/gencert/requirements/external_apps.txt
--requirement=libs.txt
#--find-links=http://pypi.pinaxproject.com
#--find-links=http://pypi2.pinaxproject.com
--find-links=http://pypi.rhizomatik.net

M2Crypto==0.20.2
pyOpenSSL==0.10
xmpppy==1.0.1
django-foafssl==0.1
python-foafcert==0.1
django-uni-form==0.6.0
django-staticfiles==0.1.2
